+++ src/script/sign.h
@@ -13,6 +13,7 @@
 #include <script/interpreter.h>
 #include <script/keyorigin.h>
 #include <script/standard.h>
+#include <script/signingprovider.h>
 #include <uint256.h>
 
 class CKey;
