+++ src/wallet/test/ismine_tests.cpp
@@ -6,6 +6,7 @@
 #include <key_io.h>
 #include <node/context.h>
 #include <script/script.h>
+#include <script/signingprovider.h>
 #include <script/standard.h>
 #include <test/util/setup_common.h>
 #include <wallet/types.h>
